Arreglado boton delete game y añadida seguridad

+++ src/main/java/controllers/developer/GameDeveloperController.java
@@ -15,8 +15,11 @@ import org.springframework.web.servlet.ModelAndView;
 
 import services.ActorService;
 import services.CategoryService;
+import services.DeveloperService;
 import services.GameService;
+import domain.Actor;
 import domain.Category;
+import domain.Developer;
 import domain.Game;
 
 @Controller
@@ -26,13 +29,16 @@ public class GameDeveloperController {
 	// Services ---------------------------------------------------------------
 
 	@Autowired
-	private GameService		gameService;
+	private GameService			gameService;
 
 	@Autowired
-	private CategoryService	categoryService;
+	private CategoryService		categoryService;
 
 	@Autowired
-	private ActorService	actorService;
+	private ActorService		actorService;
+
+	@Autowired
+	private DeveloperService	developerService;
 
 
 	// Constructors -----------------------------------------------------------
@@ -46,6 +52,14 @@ public class GameDeveloperController {
 	public ModelAndView create() {
 		ModelAndView result;
 		Game game;
+		Actor actor;
+		Developer developer;
+
+		actor = this.actorService.findByPrincipal();
+		developer = this.developerService.findByUserAccountId(actor.getUserAccount().getId());
+
+		if (developer == null)
+			return result = new ModelAndView("redirect:../../welcome/index.do");
 
 		game = this.gameService.create();
 
@@ -58,8 +72,15 @@ public class GameDeveloperController {
 	public ModelAndView edit(@RequestParam final int gameId) {
 		ModelAndView result;
 		Game game;
-
+		Actor actor;
+		Developer developer;
 		game = this.gameService.findOne(gameId);
+
+		actor = this.actorService.findByPrincipal();
+		developer = this.developerService.findByUserAccountId(actor.getUserAccount().getId());
+
+		if (developer == null || !game.getDeveloper().equals(developer))
+			return result = new ModelAndView("redirect:../../welcome/index.do");
 
 		result = this.createEditModelAndView(game);
 
@@ -93,8 +114,16 @@ public class GameDeveloperController {
 	public ModelAndView delete(@Valid final int gameId) {
 		ModelAndView result;
 		Game game;
-
+		Actor actor;
+		Developer developer;
 		game = this.gameService.findOne(gameId);
+
+		actor = this.actorService.findByPrincipal();
+		developer = this.developerService.findByUserAccountId(actor.getUserAccount().getId());
+
+		if (developer == null || !game.getDeveloper().equals(developer))
+			return result = new ModelAndView("redirect:../../welcome/index.do");
+
 		this.gameService.delete(game);
 		result = new ModelAndView("redirect:../../game/list.do");
 
@@ -102,6 +131,8 @@ public class GameDeveloperController {
 	}
 
 	// Ancillary methods ------------------------------------------------------
+
+	//Create
 	protected ModelAndView createEditModelAndView(final Game game) {
 		ModelAndView result;
 
@@ -115,37 +146,14 @@ public class GameDeveloperController {
 
 		Collection<Category> categories;
 		categories = this.categoryService.findAll();
+		if (game.getId() != 0)
+			result = new ModelAndView("game/edit");
+		else
+			result = new ModelAndView("game/create");
 
-		result = new ModelAndView("game/edit");
 		result.addObject("game", game);
 		result.addObject("categories", categories);
 		result.addObject("requestURI", "game/developer/edit.do");
-		result.addObject("message", message);
-
-		return result;
-	}
-
-	// Ancillary methods Game ----------------------------------------------
-
-	protected ModelAndView createEditModelAndViewGame(final Game game) {
-		ModelAndView result;
-
-		result = this.createEditModelAndViewGame(game, null);
-
-		return result;
-	}
-
-	protected ModelAndView createEditModelAndViewGame(final Game game, final String message) {
-		ModelAndView result;
-		Collection<Category> categories;
-
-		categories = this.categoryService.findAll();
-
-		result = new ModelAndView("game/edit");
-		result.addObject("game", game);
-		result.addObject("editProperties", true);
-		result.addObject("requestURI", "game/developer/editCategories.do?gameId=" + game.getId());
-		result.addObject("categories", categories);
 		result.addObject("message", message);
 
 		return result;
